Increases performance and simplifies url updating
The url updating logic was a bit convoluted and had a nasty bug where navigating back didn’t actually update the url.

To address this the app will now just toggle a flag if you’re dragging the slider to adjust the hue and use internal state.  After the user finishes dragging, it will toggle the flag and update the url to keep it in sync.

This has a few additional benefits:
- Performance is greatly improved since it’s only ever using the local state.
- URL and state syncing is greatly simplified.
- No more need to debounce the url update.